Add STT as a separate module
